restructure folders for containerization
