+++ app/llama_sensei/backend/add_courses/requirements.txt
@@ -6,8 +6,6 @@ transformers==4.42.4
 fastapi[standard]
 deepgram-sdk
 python-dotenv
-requests
 yt-dlp
-aiofiles
 python-dotenv
 yt-dlp-youtube-oauth2 @ https://github.com/coletdjnz/yt-dlp-youtube-oauth2/archive/refs/heads/master.zip
