+++ app/llama_sensei/backend/add_courses/vectordb/document_processor.py
@@ -1,6 +1,6 @@
-from ..embedding.get_embedding import Embedder
-from ..embedding.load_text import TranscriptLoader
-from ..embedding.preprocessing_text import TextPreprocessor
+from .get_embedding import Embedder
+from .load_text import TranscriptLoader
+from .preprocessing_text import TextPreprocessor
 from .vector_db_operations import VectorDBOperations
