+++ app/llama_sensei/backend/qa/main.py
@@ -10,14 +10,12 @@ app = FastAPI()
 
 @app.get("/")
 async def root():
-    return {"message": "Welcome to Llama Sensei API"}
+    return {"message": "LlamaSensei: Chat API"}
 
 
 @app.post("/generate_answer", response_model=ChatResponse)
 async def api_generate_answer(question: Question):
     query = GenerateRAGAnswer(question.question, question.course)
-    print((question.question, question.course))
-    # time.sleep(20)
     answer, evidence = query.generate_answer()
     return ChatResponse(answer=answer, evidence=evidence)
 
@@ -26,9 +24,9 @@ def main():
     # Run web server with uvicorn
     uvicorn.run(
         "main:app",
-        host=os.getenv("GEN_ANSWER_FASTAPI_HOST", "127.0.0.1"),
-        port=int(os.getenv("GEN_ANSWER_FASTAPI_PORT", 8000)),
-        reload=True,  # Uncomment this for debug
+        host=os.getenv("CHAT_FASTAPI_HOST", "127.0.0.1"),
+        port=int(os.getenv("CHAT_FASTAPI_PORT", 8001)),
+        # reload=True,  # Uncomment this for debug
         # workers=2,
     )
