+++ tests/unit/test_vector_db_operations.py
@@ -1,7 +1,7 @@
 import pytest
 
-from llama_sensei.backend.add_courses.embedding.get_embedding import Embedder
-from llama_sensei.backend.add_courses.embedding.preprocessing_text import TextPreprocessor
+from llama_sensei.backend.add_courses.vectordb.get_embedding import Embedder
+from llama_sensei.backend.add_courses.vectordb.preprocessing_text import TextPreprocessor
 
 text_processor = TextPreprocessor()
 embedder = Embedder()
